Booking execution moved to new namespaces
