+++ Api/Services/Accommodations/Bookings/BookingExecution/Flows/IBankCreditCardBookingFlow.cs
@@ -3,7 +3,7 @@ using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Models.Bookings;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.Flows
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.BookingExecution.Flows
 {
     public interface IBankCreditCardBookingFlow
     {
